beetime: move script to folder for publishing
To publish on AnkiWeb, the code needs to live in a folder, which is
considered a module (by virtue of there being an __init__.py file inside
of it) and which needs to be imported by one top-level script. The
top-level script and folder are to be contained in a zip file.