lookup: change criteria for due cards
Looking at the due column instead of the type also counts cards which are
due, but still in learning.  I suppose this is the way of least surprise
for users since it matches the numbers in the Decks view.  This way overdue
cards are definitely also counted.

For reference, type = 1 is learning and type = 2 is due.