settings: add a rudimentary help text dialog
This took way longer than I wanted and for all its length is still a bit
too hacky. Resizing the gridLayout container widget should resize the
textBrowser contents along with it. A close button at the bottom would
be nice. But I wanted to make a manual layout without the Qt Creator
program.