beetime: wire up the ankiweb button
Wiring this up not only took hours, literally, I was left with a bug
where at shutdown my configuration was not being saved. Some hours of
debugging later I figured out that setMod() probably sets a flag that
some other process looks at to periodically save the db, but which at
shutdown is not running anymore, so we should explicitely call save().